Update README and apply some fixes

+++ src/main/java/dev/osyx/restserver/controller/ProductService.java
@@ -72,7 +72,7 @@ public class ProductService {
         return getProductPageFromList(pageable, allProducts);
     }
 
-    public Page<Product> getProductsByPriceBetween(int minPrice, int maxPrice, Pageable paging) {
+    public Page<Product> getProductsByPriceBetween(double minPrice, double maxPrice, Pageable paging) {
         return repository.findAllByPriceBetween(minPrice, maxPrice, paging);
     }
